Birthday GIFt – Interactive Birthday Card
This is a fun and colorful interactive Birthday GIFt website made for celebrating birthday. It includes animated GIFs, clickable gifts, and a cheerful layout with a gradient background. Users can hover over gifts to see surprises and enjoy a personalized birthday message. Built with HTML and CSS, it’s a simple yet engaging project for learning web development basics.